Refactor polymer generator tests
